#391: Add family filter to Clients page
